Move stich_evaluator to utils/sticher.py + cleanup
