+++ torch_brain/utils/stitcher.py
@@ -1,5 +1,6 @@
 import logging
 from collections import defaultdict
+from typing import Callable, Iterable
 
 import hydra
 import numpy as np
@@ -7,6 +8,7 @@ import pandas as pd
 from rich import print as rprint
 import torch
 import lightning as L
+import torchmetrics
 import wandb
 
 import torch_brain
@@ -71,7 +73,151 @@ def stitch(timestamps: torch.Tensor, values: torch.Tensor) -> torch.Tensor:
     return averages
 
 
-class StitchEvaluator(L.Callback):
+class MultiSessionDecodingStitchEvaluator(L.Callback):
+    r"""A convenient stitching and evaluation framework to use when:
+     1. Your model outputs have associated timestamps
+     2. And your sampling strategy involves overlapping time windows, requiring
+        stitching to coalesce the predictions and targets before computing the
+        evaluation metric.
+     3. (Optional) You are training on multiple sessions/recordings and want to
+        compute metrics for each session individually.
+
+    This callback handles the stitching of the predictions and targets for each
+    session and computes the metric for each session. The average metric value
+    across all sessions is also computed and logged.
+    Since the stitching is done only on tensors on the same GPU, sequences that are
+    split across multiple GPUs will not be stitched together. In this case, it is
+    recommended to use a stitching-aware sampler like
+    `DistributedStitchingFixedWindowSampler` which ensures that the sequences are
+    split across GPUs in a way that allows for correct stitching.
+
+    This callback is called _after_ the validation_step, and has two main inputs:
+    - The output of the validation_step function, which is expected to be the
+      model predictions. These should be a :class:`~torch.Tensor` of shape (B, N, ...),
+      where B is the batch size, N is the number of timestamps.
+    - The batch dictionary that should have atleast the following keys:
+        - "target_values": :class:`~torch.Tensor` of shape (B, N, ...)
+        - "output_timestamps": :class:`~torch.Tensor` of shape (B, N) and dtype float
+        - "output_mask": :class:`~torch.Tensor` of shape (B, N) and dtype bool
+        - "session_id": A list of session IDs for each sample in the batch
+        - "absolute_start": :class:`~torch.Tensor` of shape (B) and dtype float
+
+    This callback operates by maintaining a cache of the predictions, targets, and
+    timestamps for each session. This cache is updated at the end of each batch.
+    Finally, at the end of the epoch, the cache is coalesced and the metric is
+    computed for each session.
+
+    Note:
+        This callback is designed to work only with metric functions from
+         the `torchmetrics` library.
+    """
+
+    def __init__(
+        self,
+        session_ids: Iterable[str],
+        metric_factory: Callable[[], torchmetrics.Metric],
+        quiet=False,
+    ):
+        r"""
+        Args:
+            session_ids: An iterable of session IDs for which the metrics are to be computed.
+            torchmetric_factory: A callable that returns an instance of a `torchmetrics.Metric`.
+            quiet: If True, disables the logging of the metrics to the console.
+        """
+        self.quiet = quiet
+
+        if not isinstance(metric_factory(), torchmetrics.Metric):
+            raise ValueError(
+                "`metric_factory()` must return an instance of torchmetrics.Metric."
+            )
+
+        self.metrics = {k: metric_factory() for k in session_ids}
+
+    def on_validation_epoch_start(self, trainer, pl_module):
+        # Cache to store the predictions, targets, and timestamps for each
+        # validation step. This will be coalesced at the end of the validation,
+        # using the stitch function.
+        self.cache = defaultdict(
+            lambda: {
+                "pred": [],
+                "target": [],
+                "timestamps": [],
+            }
+        )
+
+    def on_validation_batch_end(self, trainer, pl_module, outputs, batch, batch_idx):
+        # Update the cache with the predictions, targets, timestamps, and subtask index
+        # TODO: Mask timestamps based on subtask_index
+        batch_size = len(outputs)
+        for i in range(batch_size):
+            mask = batch["output_mask"][i]
+            session_id = batch["session_id"][i]
+            absolute_start = batch["absolute_start"][i]
+
+            pred = outputs[i][mask]
+            target = batch["target_values"][i][mask]
+            timestamps = batch["output_timestamps"][i][mask] + absolute_start
+
+            self.cache[session_id]["pred"].append(pred.detach())
+            self.cache[session_id]["target"].append(target.detach())
+            self.cache[session_id]["timestamps"].append(timestamps.detach())
+
+    def on_validation_epoch_end(self, trainer, pl_module, prefix="val"):
+        # compute metric for each session
+        metrics = {}
+        for session_id, metric_fn in self.metrics.items():
+            cache = self.cache[session_id]
+            pred = torch.cat(cache["pred"])
+            target = torch.cat(cache["target"])
+            timestamps = torch.cat(cache["timestamps"])
+
+            stitched_pred = stitch(timestamps, pred)
+            stitched_target = stitch(timestamps, target)
+
+            metric_fn.to(pl_module.device).update(stitched_pred, stitched_target)
+            metrics[session_id] = metric_fn.compute()
+            metric_fn.reset()
+
+        # compute the average metric
+        metrics[f"average_{prefix}_metric"] = torch.tensor(
+            list(metrics.values())
+        ).mean()
+
+        # log the metrics
+        self.log_dict(metrics)
+        if not self.quiet:
+            logging.info(f"Logged {len(metrics)} {prefix} metrics.")
+
+        metrics_data = []
+        for metric_name, metric_value in metrics.items():
+            metrics_data.append({"metric": metric_name, "value": metric_value.item()})
+
+        metrics_df = pd.DataFrame(metrics_data)
+        if not self.quiet:
+            rprint(metrics_df)
+
+        if trainer.is_global_zero:
+            for logger in trainer.loggers:
+                if isinstance(logger, L.pytorch.loggers.TensorBoardLogger):
+                    logger.experiment.add_text(
+                        f"{prefix}_metrics", metrics_df.to_markdown()
+                    )
+                if isinstance(logger, L.pytorch.loggers.WandbLogger):
+                    logger.experiment.log(
+                        {f"{prefix}_metrics": wandb.Table(dataframe=metrics_df)}
+                    )
+
+    def on_test_epoch_start(self, *args, **kwargs):
+        self.on_validation_epoch_start(*args, **kwargs)
+
+    def on_test_batch_end(self, *args, **kwargs):
+        self.on_validation_batch_end(*args, **kwargs)
+
+    def on_test_epoch_end(self, *args, **kwargs):
+        self.on_validation_epoch_end(*args, **kwargs, prefix="test")
+
+
+class MultiSessionMultiTaskStitchEvaluator(L.Callback):
     def __init__(self, dataset_config_dict: dict):
         metrics = defaultdict(lambda: defaultdict(dict))
         # setup the metrics
